requirements.txt
pandas==2.2.3
playwright==1.51.0
python-dotenv==1.1.0
Requests==2.32.3
